# workload/tpcc: re-order NewOrder txn locking