Informs #131585.

This commit updates the `NewOrder` transaction to read from the `warehouse` table before updating the `district` table. This ensures a consistent lock ordering with the `Payment` transaction, which updates both tables. A consistent lock ordering for both reads and writes is necessary to avoid transaction deadlocks under Serializable isolation.

Release note: None