@nvanbenschoten 2024-11-11
I ran TPC-C nowait (`workload run tpcc --wait=false --isolation-level=serializable --conn-vars=enable_durable_locking_for_serializable=true --duration 5m0s --ramp 1m0s --warehouses 5 {pgurl:1-3}`) with and without this change. As expected, I see a significant reduction in transaction deadlocks, combined with a small increase in throughput:

<img width="888" alt="Screenshot 2024-11-11 at 4 42 25 PM" src="https://github.com/user-attachments/assets/b404640d-d648-451f-8a63-503081aea62a">
<img width="887" alt="Screenshot 2024-11-11 at 4 42 57 PM" src="https://github.com/user-attachments/assets/1488820b-dfae-4519-a924-222f58771233">

### Before

```
_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
  300.0s        0           4323           14.4    107.3     83.9    260.0    352.3    805.3  delivery

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
  300.0s        0          43207          144.0    127.6     96.5    335.5    503.3   1811.9  newOrder

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
  300.0s        0           4320           14.4     41.8     10.0    184.5    285.2    604.0  orderStatus

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
  300.0s        0          43187          144.0    198.9    176.2    469.8    738.2   1946.2  payment

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
  300.0s        0           4313           14.4     58.6     27.3    209.7    335.5    503.3  stockLevel

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__result
  300.0s        0          99350          331.2    151.0    125.8    402.7    637.5   1946.2
```
### After
```
_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
  300.0s        0           4586           15.3    159.7    134.2    335.5    469.8    805.3  delivery

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
  300.0s        0          45835          152.8    108.5     88.1    251.7    369.1   1073.7  newOrder

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
  300.0s        0           4579           15.3     38.9     13.6    151.0    218.1    369.1  orderStatus

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
  300.0s        0          45834          152.8    193.6    184.5    352.3    486.5   1342.2  payment

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
  300.0s        0           4583           15.3     53.6     29.4    167.8    243.3    486.5  stockLevel

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__result
  300.0s        0         105417          351.4    142.3    130.0    318.8    436.2   1342.2
```

----

The next question is "why are we still seeing any deadlocks?". Offline, I had hypothesized that this was due to stateful transactions retries (through `savepoint cockroach_restart`) causing a lock ordering violation deadlock. To test this hypothesis, I ran with `--txn-retries=false --tolerate-errors=true`, which disables stateful txn retries. With this change, transaction deadlocks dropped to zero.
<img width="899" alt="Screenshot 2024-11-11 at 4 58 37 PM" src="https://github.com/user-attachments/assets/2c2d3667-a09c-4c99-92ff-edea322e5625">
<img width="888" alt="Screenshot 2024-11-11 at 4 58 43 PM" src="https://github.com/user-attachments/assets/5d140552-c420-46f3-a3bb-f2fdbbccbdef">